Events commands structural refactor
