+++ src/Bot/Commands/DiscordClientExtensions_Commands.cs
@@ -1,5 +1,6 @@
 ﻿using DSharpPlus;
 using DSharpPlus.CommandsNext;
+using Ivao.It.DiscordBot.Commands.EventsWorkflow;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Ivao.It.DiscordBot.Commands;
@@ -15,9 +16,10 @@ internal static class DiscordClientExtensions_Commands
         });
 
         //Commands
+        //commands.RegisterCommands(Assembly.GetExecutingAssembly()); //Events are not public.
         commands.RegisterCommands<BotCommands>();
-        commands.RegisterCommands<EventsWorkflowCommands>();
-
+        commands.RegisterCommands<EventsCommands>();
+        
         //Global Event Handlers
         commands.CommandExecuted += CommandsNextEventHandlers.Commands_CommandExecuted;
         commands.CommandErrored += CommandsNextEventHandlers.Commands_CommandErrored;
